+++ odoo/local-src/specific_fct/models/__init__.py
@@ -4,10 +4,10 @@ from . import product
 from . import sale
 from . import stock
 from . import partner
-from . import account_move
 from . import account_invoice
 from . import account_payment_mode
 from . import account_tax
 from . import purchase
 from . import stock_batch_picking
 from . import ir_qweb
+from . import ir_attachment