+++ odoo/local-src/specific_fct/models/partner.py
@@ -3,7 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 import logging
 
-from openerp import api, fields, models, SUPERUSER_ID
+from openerp import api, fields, models
 
 _logger = logging.getLogger(__name__)
 
@@ -95,61 +95,3 @@ class ResPartner(models.Model):
             parts = (p.strip() for p in name.splitlines() if p.strip())
             res.append((partner_id, ', '.join(parts)))
         return res
-
-    @api.model
-    def _trgm_extension_exists(self):
-        self.env.cr.execute("""
-            SELECT name, installed_version
-            FROM pg_available_extensions
-            WHERE name = 'pg_trgm'
-            LIMIT 1;
-            """)
-
-        extension = self.env.cr.fetchone()
-        if extension is None:
-            return 'missing'
-
-        if extension[1] is None:
-            return 'uninstalled'
-
-        return 'installed'
-
-    @api.model
-    def _is_postgres_superuser(self):
-        self.env.cr.execute("SHOW is_superuser;")
-        superuser = self.env.cr.fetchone()
-        return superuser is not None and superuser[0] == 'on' or False
-
-    @api.model
-    def _install_trgm_extension(self):
-        extension = self._trgm_extension_exists()
-        if extension == 'missing':
-            _logger.warning('To use pg_trgm you have to install the '
-                            'postgres-contrib module.')
-        elif extension == 'uninstalled':
-            if self._is_postgres_superuser():
-                self.env.cr.execute("CREATE EXTENSION IF NOT EXISTS pg_trgm;")
-                return True
-            else:
-                _logger.warning('To use pg_trgm you have to create the '
-                                'extension pg_trgm in your database or you '
-                                'have to be the superuser.')
-        else:
-            return True
-        return False
-
-    def init(self, cr):
-        installed = self._install_trgm_extension(cr, SUPERUSER_ID, context={})
-        cr.commit()
-
-        def create_gin_trgm_index(field):
-            index_name = 'res_partner_%s_trgm_index' % field
-            cr.execute('SELECT indexname FROM pg_indexes WHERE indexname = %s',
-                       (index_name,))
-            if not cr.fetchone():
-                cr.execute('CREATE INDEX %s '
-                           'ON res_partner '
-                           'USING gin (%s gin_trgm_ops)' % (index_name, field))
-        if installed:
-            for field in ('email', 'display_name'):
-                create_gin_trgm_index(field)