+++ odoo/local-src/speedy_views/models/account_invoice.py
@@ -2,70 +2,35 @@
 # Copyright 2016 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
-import logging
 
-from openerp import models, api, SUPERUSER_ID
-
-_logger = logging.getLogger(__name__)
+from openerp import fields, models, api, SUPERUSER_ID
+from .utils import install_trgm_extension, create_index
 
 
 class AccountInvoice(models.Model):
     _inherit = 'account.invoice'
 
-    @api.model
-    def _trgm_extension_exists(self):
-        self.env.cr.execute("""
-            SELECT name, installed_version
-            FROM pg_available_extensions
-            WHERE name = 'pg_trgm'
-            LIMIT 1;
-            """)
-
-        extension = self.env.cr.fetchone()
-        if extension is None:
-            return 'missing'
-
-        if extension[1] is None:
-            return 'uninstalled'
-
-        return 'installed'
-
-    @api.model
-    def _is_postgres_superuser(self):
-        self.env.cr.execute("SHOW is_superuser;")
-        superuser = self.env.cr.fetchone()
-        return superuser is not None and superuser[0] == 'on' or False
-
-    @api.model
-    def _install_trgm_extension(self):
-        extension = self._trgm_extension_exists()
-        if extension == 'missing':
-            _logger.warning('To use pg_trgm you have to install the '
-                            'postgres-contrib module.')
-        elif extension == 'uninstalled':
-            if self._is_postgres_superuser():
-                self.env.cr.execute("CREATE EXTENSION IF NOT EXISTS pg_trgm;")
-                return True
-            else:
-                _logger.warning('To use pg_trgm you have to create the '
-                                'extension pg_trgm in your database or you '
-                                'have to be the superuser.')
-        else:
-            return True
-        return False
+    partner_id = fields.Many2one(index=True)
+    claim_id = fields.Many2one(index=True)
 
     def init(self, cr):
-        installed = self._install_trgm_extension(cr, SUPERUSER_ID, context={})
+        env = api.Environment(cr, SUPERUSER_ID, {})
+        trgm_installed = install_trgm_extension(env)
         cr.commit()
-        if installed:
-            cr.execute("""
-                SELECT indexname
-                FROM pg_indexes
-                WHERE indexname = 'account_invoice_origin_gin_trgm'
-            """)
-            if not cr.fetchone():
-                cr.execute("""
-                    CREATE INDEX account_invoice_origin_gin_trgm
-                    ON account_invoice
-                    USING gin (origin gin_trgm_ops)
-                """)
+
+        if trgm_installed:
+            index_name = 'account_invoice_origin_gin_trgm'
+            create_index(cr, index_name, self._table,
+                         'USING gin (origin gin_trgm_ops)')
+
+        # default list view sort by those fields desc
+        index_name = 'account_invoice_list_sort_index'
+        create_index(cr, index_name, self._table,
+                     '(date_invoice desc, number desc, id desc) '
+                     'WHERE active')
+
+        # active is mostly used with 'true' so this partial index improves
+        # globally the queries. The same index on (active) without where
+        # would in general not be used.
+        index_name = 'account_invoice_active_true_index'
+        create_index(cr, index_name, self._table, '(active) where active')